requirements.txt
aiogram==3.13.1
aiomysql==0.3.2
openai==1.54.5
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.32.0
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

